Wrap cstg with feature flag
