+++ features.json
@@ -1,4 +1,5 @@
 [
   "NATIVE",
-  "VIDEO"
+  "VIDEO",
+  "UID2_CSTG"
 ]
