+++ package.json
@@ -131,7 +131,6 @@
     "dset": "3.1.2",
     "express": "^4.15.4",
     "fun-hooks": "^0.9.9",
-    "gulp-cli": "^2.3.0",
     "just-clone": "^1.0.2",
     "live-connect-js": "^6.0.0"
   },
